test: fixing e2e test failure due to volume in use error (#55)
Issue #, if available:
261
*Description of changes:*
Finch Container Development E2E Tests Compose down command fails to
delete the volumes due to volume being used by compose up command. This
PR added a delay of 10 sec after the compose up command to allow it to
fully up and running the service. This additional wait will prevent the
error of not being able to delete the volume by compose down command due
to concurrent access.

The compose down command fails to delete the volumes because the compose
up command is still using them. This PR adds a 10-second delay after the
compose up command to avoid the error of not being able to delete the
volume by the compose down command due to simultaneous access.

*Testing done:*
Yes


- [x] I've reviewed the guidance in CONTRIBUTING.md


#### License Acceptance

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

Signed-off-by: Monirul Islam <monirulu@amazon.com>