fix: Retry image pull for 3 times and then fail (#137)
Issue #, if available:

*Description of changes:*
Seems the ecr connection is reset during runtime which restricts any
chance for successful pull for open connections. Adding retry to see if
that can solve the issue. A more broad approach can to retry the entire
process.

*Testing done:*



- [ ] I've reviewed the guidance in CONTRIBUTING.md


#### License Acceptance

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

Signed-off-by: Shubharanshu Mahapatra <shubhum@amazon.com>